+++ App/Controllers/PrispevkyController.php
@@ -4,6 +4,7 @@ namespace App\Controllers;
 
 use App\Core\AControllerBase;
 use App\Core\Responses\Response;
+use App\Models\Comment;
 use App\Models\Prispevok;
 
 class PrispevkyController extends AControllerBase
@@ -15,7 +16,7 @@ class PrispevkyController extends AControllerBase
      */
     public function authorize($action)
     {
-        switch ($action){
+        switch ($action) {
             case "delete":
             case "store":
             case "create":
@@ -31,41 +32,75 @@ class PrispevkyController extends AControllerBase
         return $this->html($prispevky);
     }
 
-    public function delete() {
-
+    public function delete()
+    {
+        $data = Prispevok::getAll();
+        $commdata = Comment::getAll();
         $id = $this->request()->getValue('id');
-        $postToDelete = Prispevok::getOne($id);
-        if ($postToDelete) {
-            $postToDelete->delete();
+        $i = false;
+
+        foreach ($data as $log) {
+            if ($log->getIdPrispevku() == $id) {
+                $i = true;
+                break;
+            }
         }
+
+        if ($i) {
+            $id = $this->request()->getValue('id');
+            $postToDelete = Prispevok::getOne($id);
+
+            foreach ($commdata as $comm) {
+                if ($comm->getCommentPost() == $id) {
+                    $commId = $comm->getCommentId();
+                    $commenttoDelete = Comment::getOne($commId);
+                    if ($commenttoDelete) {
+                        $commenttoDelete->delete();
+                    }
+                }
+            }
+
+            if ($postToDelete) {
+                $postToDelete->delete();
+            }
+
         return $this->redirect("?c=prispevky");
+        }
     }
 
-    public function store() {
+    public function store()
+    {
         $id = $this->request()->getValue('id');
-        $post = ( $id ? Prispevok::getOne($id) : new Prispevok());
+        $post = ($id ? Prispevok::getOne($id) : new Prispevok());
         $error = 0;
 
         if ($this->request()->getValue('obrazok') == null) {
             $post->setObrazok("Nezadany");
             $error++;
+        } else {
+            $post->setObrazok($this->request()->getValue('obrazok'));
         }
-        else{$post->setObrazok($this->request()->getValue('obrazok'));}
 
         if ($this->request()->getValue('nazov') == null) {
             $post->setNazov("Nezadany");
             $error++;
-        } else {$post->setNazov($this->request()->getValue('nazov'));}
+        } else {
+            $post->setNazov($this->request()->getValue('nazov'));
+        }
 
         if ($this->request()->getValue('text') == null) {
             $post->setText("Nezadany");
             $error++;
-        } else{$post->setText($this->request()->getValue('text'));}
+        } else {
+            $post->setText($this->request()->getValue('text'));
+        }
 
-        if ($error > 0) {return $this->redirect("?c=prispevky&a=create");
+        if ($error > 0) {
+            return $this->redirect("?c=prispevky&a=create");
         } else {
             $post->save();
-            return $this->redirect("?c=prispevky");}
+            return $this->redirect("?c=prispevky");
+        }
     }
 
     public function create() {
@@ -73,8 +108,20 @@ class PrispevkyController extends AControllerBase
     }
 
     public function edit() {
+        $data = Prispevok::getAll();
         $id = $this->request()->getValue('id');
-        $postToEdit = Prispevok::getOne($id);
-        return $this->html($postToEdit, viewName: 'create.form');
+        $i = false;
+
+        foreach ($data as $log) {
+            if ($log->getIdPrispevku() == $id) {
+                $i = true;
+                break;
+            }
+        }
+
+        if ($i) {
+            $postToEdit = Prispevok::getOne($id);
+            return $this->html($postToEdit, viewName: 'create.form');
+        } else {return $this->redirect("?c=prispevky");}
     }
 }
